🐛 fix: serialize pause/resume status writes (code review)
Addresses two review findings on the producer commit:

- pauseSimulation/resumeSimulation spawned independent unstructured Tasks, so a
  rapid pause→resume could commit `.paused` AFTER `.running`, parking a live run
  as `.paused` in the DB (a real gap once PR2's auto-pause-on-tab-switch makes
  rapid pairs common). Route both through a serialized statusWriteTask chain so
  writes commit in call order; the terminal ladder drains it before deciding.
- Reword the persistCheckpoint doc: checkpoint vs status write is safe due to
  DatabaseWriter's serialized write transactions + in-transaction fetch (full-row
  read-modify-write preserves the other column), not "orthogonal columns".

Resume/checkpoint tests moved to a sibling SimulationViewModelStatusTests+Resume
extension (file_length cap; same suite, stays serialized).

Part of #656
